+++ core-tests/src/test/java/org/mule/runtime/core/processor/AsyncDelegateMessageProcessorTestCase.java
@@ -26,8 +26,8 @@ import org.mule.runtime.api.exception.MuleException;
 import org.mule.runtime.core.api.context.WorkManager;
 import org.mule.runtime.core.api.context.WorkManagerSource;
 import org.mule.runtime.core.api.processor.Processor;
-import org.mule.runtime.core.api.processor.factory.AsynchronousProcessingStrategyFactory;
 import org.mule.runtime.core.api.routing.RoutingException;
+import org.mule.runtime.core.api.strategy.factory.AsynchronousProcessingStrategyFactory;
 import org.mule.runtime.core.api.transaction.Transaction;
 import org.mule.runtime.core.construct.Flow;
 import org.mule.runtime.core.transaction.TransactionCoordination;
