+++ core/src/main/java/org/mule/runtime/core/api/construct/Pipeline.java
@@ -10,8 +10,8 @@ import org.mule.runtime.core.api.processor.MessageProcessorContainer;
 import org.mule.runtime.core.api.processor.ProcessingDescriptor;
 import org.mule.runtime.core.api.processor.ProcessingStrategy;
 import org.mule.runtime.core.api.processor.Processor;
-import org.mule.runtime.core.api.processor.factory.ProcessingStrategyFactory;
 import org.mule.runtime.core.api.source.MessageSource;
+import org.mule.runtime.core.api.strategy.factory.ProcessingStrategyFactory;
 
 import java.util.List;
 
@@ -30,8 +30,6 @@ public interface Pipeline extends FlowConstruct, MessageProcessorContainer, Proc
   public List<Processor> getMessageProcessors();
 
   public void setProcessingStrategyFactory(ProcessingStrategyFactory processingStrategyFactory);
-
-  public ProcessingStrategyFactory getProcessingStrategyFactory();
 
   public ProcessingStrategy getProcessingStrategy();
