--- core/src/main/java/org/mule/runtime/core/api/processor/factory/SynchronousProcessingStrategyFactory.java
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
- * The software in this package is published under the terms of the CPAL v1.0
- * license, a copy of which has been included with this distribution in the
- * LICENSE.txt file.
- */
-package org.mule.runtime.core.api.processor.factory;
-
-import org.mule.runtime.core.api.processor.ProcessingStrategy;
-import org.mule.runtime.core.processor.strategy.SynchronousProcessingStrategy;
-
-
-public class SynchronousProcessingStrategyFactory implements ProcessingStrategyFactory {
-
-  @Override
-  public ProcessingStrategy create() {
-    return new SynchronousProcessingStrategy();
-  }
-
-}
