+++ core/src/main/resources/META-INF/mule-module.properties
@@ -29,7 +29,6 @@ artifact.export.classPackages=org.mule.runtime.core,\
                               org.mule.runtime.core.api.model,\
                               org.mule.runtime.core.api.object,\
                               org.mule.runtime.core.api.processor,\
-                              org.mule.runtime.core.api.processor.factory,\
                               org.mule.runtime.core.api.registry,\
                               org.mule.runtime.core.api.retry,\
                               org.mule.runtime.core.api.routing,\
@@ -42,6 +41,7 @@ artifact.export.classPackages=org.mule.runtime.core,\
                               org.mule.runtime.core.api.service,\
                               org.mule.runtime.core.api.source,\
                               org.mule.runtime.core.api.store,\
+                              org.mule.runtime.core.api.strategy.factory,\
                               org.mule.runtime.core.api.transaction,\
                               org.mule.runtime.core.api.transformer,\
                               org.mule.runtime.core.api.transformer.wire,\
