+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/handlers/MuleNamespaceHandler.java
@@ -87,9 +87,9 @@ import org.mule.runtime.config.spring.util.SpringBeanLookup;
 import org.mule.runtime.core.api.config.MuleProperties;
 import org.mule.runtime.core.api.config.ThreadingProfile;
 import org.mule.runtime.core.api.processor.LoggerMessageProcessor;
-import org.mule.runtime.core.api.processor.factory.AsynchronousProcessingStrategyFactory;
-import org.mule.runtime.core.api.processor.factory.NonBlockingProcessingStrategyFactory;
 import org.mule.runtime.core.api.source.MessageSource;
+import org.mule.runtime.core.api.strategy.factory.AsynchronousProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.NonBlockingProcessingStrategyFactory;
 import org.mule.runtime.core.component.DefaultJavaComponent;
 import org.mule.runtime.core.component.PooledJavaComponent;
 import org.mule.runtime.core.component.SimpleCallableJavaComponent;
