+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/util/ProcessingStrategyParserUtils.java
@@ -10,7 +10,7 @@ import static org.mule.runtime.config.spring.dsl.model.ApplicationModel.PROCESSI
 import static org.mule.runtime.config.spring.dsl.model.ApplicationModel.PROCESSING_STRATEGY_FACTORY_ATTRIBUTE;
 import static org.mule.runtime.core.util.ProcessingStrategyUtils.parseProcessingStrategy;
 
-import org.mule.runtime.core.api.processor.factory.ProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.ProcessingStrategyFactory;
 
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.config.RuntimeBeanReference;
