+++ tests/integration/src/test/java/org/mule/test/config/spring/flow/FlowProcessingStrategyConfigTestCase.java
@@ -12,14 +12,14 @@ import org.mule.runtime.core.api.MuleContext;
 import org.mule.runtime.core.api.processor.MessageProcessorChainBuilder;
 import org.mule.runtime.core.api.processor.ProcessingStrategy;
 import org.mule.runtime.core.api.processor.Processor;
-import org.mule.runtime.core.api.processor.factory.ProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.AsynchronousProcessingStrategyFactory.AsynchronousProcessingStrategy;
+import org.mule.runtime.core.api.strategy.factory.DefaultFlowProcessingStrategyFactory.DefaultFlowProcessingStrategy;
+import org.mule.runtime.core.api.strategy.factory.NonBlockingProcessingStrategyFactory.NonBlockingProcessingStrategy;
+import org.mule.runtime.core.api.strategy.factory.ProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.SynchronousProcessingStrategyFactory.SynchronousProcessingStrategy;
 import org.mule.runtime.core.construct.Flow;
-import org.mule.runtime.core.construct.flow.DefaultFlowProcessingStrategy;
 import org.mule.runtime.core.processor.AsyncDelegateMessageProcessor;
 import org.mule.runtime.core.processor.strategy.AbstractThreadingProfileProcessingStrategy;
-import org.mule.runtime.core.processor.strategy.AsynchronousProcessingStrategy;
-import org.mule.runtime.core.processor.strategy.NonBlockingProcessingStrategy;
-import org.mule.runtime.core.processor.strategy.SynchronousProcessingStrategy;
 import org.mule.test.AbstractIntegrationTestCase;
 
 import java.util.List;
