+++ tests/integration/src/test/java/org/mule/test/core/NonBlockingFullySupportedFunctionalTestCase.java
@@ -15,9 +15,9 @@ import static org.mule.functional.functional.FlowAssert.verify;
 import org.mule.runtime.core.MessageExchangePattern;
 import org.mule.runtime.core.api.Event;
 import org.mule.runtime.core.api.context.MuleContextBuilder;
-import org.mule.runtime.core.api.processor.factory.DefaultFlowProcessingStrategyFactory;
-import org.mule.runtime.core.api.processor.factory.NonBlockingProcessingStrategyFactory;
-import org.mule.runtime.core.api.processor.factory.ProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.DefaultFlowProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.NonBlockingProcessingStrategyFactory;
+import org.mule.runtime.core.api.strategy.factory.ProcessingStrategyFactory;
 import org.mule.runtime.core.config.DefaultMuleConfiguration;
 import org.mule.test.AbstractIntegrationTestCase;
 import org.mule.test.runner.RunnerDelegateTo;
